soroban_env_common
Module meta

This module contains version constants (and other metadata) that are embedded in binaries (especially WASM binaries) compiled against a particular version of this crate. Versioning at this level provides an early diagnostic check for compatibility between a loaded WASM binary and the Env interface provided by the host, rather than a cryptic failure due to a runtime host function signature mismatch.

Constants

ENV_META_V0_SECTION_NAME

INTERFACE_VERSION

XDR

Functions

get_ledger_protocol_version

get_pre_release_version

make_interface_version