soroban_env_common
Module xdr

Modules

int128_helpers

Structs

AccountEntry

AccountEntry is an XDR Struct defines as:

AccountEntryExtensionV1

AccountEntryExtensionV1 is an XDR Struct defines as:

AccountEntryExtensionV2

AccountEntryExtensionV2 is an XDR Struct defines as:

AccountEntryExtensionV3

AccountEntryExtensionV3 is an XDR Struct defines as:

AccountId

AccountId is an XDR Typedef defines as:

AllowTrustOp

AllowTrustOp is an XDR Struct defines as:

AlphaNum4

AlphaNum4 is an XDR Struct defines as:

AlphaNum12

AlphaNum12 is an XDR Struct defines as:

AssetCode4

AssetCode4 is an XDR Typedef defines as:

AssetCode12

AssetCode12 is an XDR Typedef defines as:

Auth

Auth is an XDR Struct defines as:

AuthCert

AuthCert is an XDR Struct defines as:

AuthenticatedMessageV0

AuthenticatedMessageV0 is an XDR NestedStruct defines as:

BeginSponsoringFutureReservesOp

BeginSponsoringFutureReservesOp is an XDR Struct defines as:

BucketMetadata

BucketMetadata is an XDR Struct defines as:

BumpSequenceOp

BumpSequenceOp is an XDR Struct defines as:

BytesM

ChangeTrustOp

ChangeTrustOp is an XDR Struct defines as:

ClaimClaimableBalanceOp

ClaimClaimableBalanceOp is an XDR Struct defines as:

ClaimLiquidityAtom

ClaimLiquidityAtom is an XDR Struct defines as:

ClaimOfferAtom

ClaimOfferAtom is an XDR Struct defines as:

ClaimOfferAtomV0

ClaimOfferAtomV0 is an XDR Struct defines as:

ClaimableBalanceEntry

ClaimableBalanceEntry is an XDR Struct defines as:

ClaimableBalanceEntryExtensionV1

ClaimableBalanceEntryExtensionV1 is an XDR Struct defines as:

ClaimantV0

ClaimantV0 is an XDR NestedStruct defines as:

ClawbackClaimableBalanceOp

ClawbackClaimableBalanceOp is an XDR Struct defines as:

ClawbackOp

ClawbackOp is an XDR Struct defines as:

ConfigSettingContractBandwidthV0

ConfigSettingContractBandwidthV0 is an XDR Struct defines as:

ConfigSettingContractComputeV0

ConfigSettingContractComputeV0 is an XDR Struct defines as:

ConfigSettingContractEventsV0

ConfigSettingContractEventsV0 is an XDR Struct defines as:

ConfigSettingContractExecutionLanesV0

ConfigSettingContractExecutionLanesV0 is an XDR Struct defines as:

ConfigSettingContractHistoricalDataV0

ConfigSettingContractHistoricalDataV0 is an XDR Struct defines as:

ConfigSettingContractLedgerCostV0

ConfigSettingContractLedgerCostV0 is an XDR Struct defines as:

ConfigUpgradeSet

ConfigUpgradeSet is an XDR Struct defines as:

ConfigUpgradeSetKey

ConfigUpgradeSetKey is an XDR Struct defines as:

ContractCodeCostInputs

ContractCodeCostInputs is an XDR Struct defines as:

ContractCodeEntry

ContractCodeEntry is an XDR Struct defines as:

ContractCodeEntryV1

ContractCodeEntryV1 is an XDR NestedStruct defines as:

ContractCostParamEntry

ContractCostParamEntry is an XDR Struct defines as:

ContractCostParams

ContractCostParams is an XDR Typedef defines as:

ContractDataEntry

ContractDataEntry is an XDR Struct defines as:

ContractEvent

ContractEvent is an XDR Struct defines as:

ContractEventV0

ContractEventV0 is an XDR NestedStruct defines as:

ContractIdPreimageFromAddress

ContractIdPreimageFromAddress is an XDR NestedStruct defines as:

CreateAccountOp

CreateAccountOp is an XDR Struct defines as:

CreateClaimableBalanceOp

CreateClaimableBalanceOp is an XDR Struct defines as:

CreateContractArgs

CreateContractArgs is an XDR Struct defines as:

CreatePassiveSellOfferOp

CreatePassiveSellOfferOp is an XDR Struct defines as:

Curve25519Public

Curve25519Public is an XDR Struct defines as:

Curve25519Secret

Curve25519Secret is an XDR Struct defines as:

DataEntry

DataEntry is an XDR Struct defines as:

DataValue

DataValue is an XDR Typedef defines as:

DecoratedSignature

DecoratedSignature is an XDR Struct defines as:

DiagnosticEvent

DiagnosticEvent is an XDR Struct defines as:

DiagnosticEvents

DiagnosticEvents is an XDR Typedef defines as:

DontHave

DontHave is an XDR Struct defines as:

Duration

Duration is an XDR Typedef defines as:

EncryptedBody

EncryptedBody is an XDR Typedef defines as:

EvictionIterator

EvictionIterator is an XDR Struct defines as:

ExtendFootprintTtlOp

ExtendFootprintTtlOp is an XDR Struct defines as:

FeeBumpTransaction

FeeBumpTransaction is an XDR Struct defines as:

FeeBumpTransactionEnvelope

FeeBumpTransactionEnvelope is an XDR Struct defines as:

FloodAdvert

FloodAdvert is an XDR Struct defines as:

FloodDemand

FloodDemand is an XDR Struct defines as:

Frame

Hash

Hash is an XDR Typedef defines as:

HashIdPreimageContractId

HashIdPreimageContractId is an XDR NestedStruct defines as:

HashIdPreimageOperationId

HashIdPreimageOperationId is an XDR NestedStruct defines as:

HashIdPreimageRevokeId

HashIdPreimageRevokeId is an XDR NestedStruct defines as:

HashIdPreimageSorobanAuthorization

HashIdPreimageSorobanAuthorization is an XDR NestedStruct defines as:

Hello

Hello is an XDR Struct defines as:

HmacSha256Key

HmacSha256Key is an XDR Struct defines as:

HmacSha256Mac

HmacSha256Mac is an XDR Struct defines as:

InflationPayout

InflationPayout is an XDR Struct defines as:

InnerTransactionResult

InnerTransactionResult is an XDR Struct defines as:

InnerTransactionResultPair

InnerTransactionResultPair is an XDR Struct defines as:

Int128Parts

Int128Parts is an XDR Struct defines as:

Int256Parts

Int256Parts is an XDR Struct defines as:

InvokeContractArgs

InvokeContractArgs is an XDR Struct defines as:

InvokeHostFunctionOp

InvokeHostFunctionOp is an XDR Struct defines as:

InvokeHostFunctionSuccessPreImage

InvokeHostFunctionSuccessPreImage is an XDR Struct defines as:

LedgerBounds

LedgerBounds is an XDR Struct defines as:

LedgerCloseMetaExtV1

LedgerCloseMetaExtV1 is an XDR Struct defines as:

LedgerCloseMetaV0

LedgerCloseMetaV0 is an XDR Struct defines as:

LedgerCloseMetaV1

LedgerCloseMetaV1 is an XDR Struct defines as:

LedgerCloseValueSignature

LedgerCloseValueSignature is an XDR Struct defines as:

LedgerEntry

LedgerEntry is an XDR Struct defines as:

LedgerEntryChanges

LedgerEntryChanges is an XDR Typedef defines as:

LedgerEntryExtensionV1

LedgerEntryExtensionV1 is an XDR Struct defines as:

LedgerFootprint

LedgerFootprint is an XDR Struct defines as:

LedgerHeader

LedgerHeader is an XDR Struct defines as:

LedgerHeaderExtensionV1

LedgerHeaderExtensionV1 is an XDR Struct defines as:

LedgerHeaderHistoryEntry

LedgerHeaderHistoryEntry is an XDR Struct defines as:

LedgerKeyAccount

LedgerKeyAccount is an XDR NestedStruct defines as:

LedgerKeyClaimableBalance

LedgerKeyClaimableBalance is an XDR NestedStruct defines as:

LedgerKeyConfigSetting

LedgerKeyConfigSetting is an XDR NestedStruct defines as:

LedgerKeyContractCode

LedgerKeyContractCode is an XDR NestedStruct defines as:

LedgerKeyContractData

LedgerKeyContractData is an XDR NestedStruct defines as:

LedgerKeyData

LedgerKeyData is an XDR NestedStruct defines as:

LedgerKeyLiquidityPool

LedgerKeyLiquidityPool is an XDR NestedStruct defines as:

LedgerKeyOffer

LedgerKeyOffer is an XDR NestedStruct defines as:

LedgerKeyTrustLine

LedgerKeyTrustLine is an XDR NestedStruct defines as:

LedgerKeyTtl

LedgerKeyTtl is an XDR NestedStruct defines as:

LedgerScpMessages

LedgerScpMessages is an XDR Struct defines as:

Liabilities

Liabilities is an XDR Struct defines as:

Limited

Limited wraps an object and provides functions for enforcing limits.

Limits

Limits contains the limits that a limited reader or writer will be constrained to.

LiquidityPoolConstantProductParameters

LiquidityPoolConstantProductParameters is an XDR Struct defines as:

LiquidityPoolDepositOp

LiquidityPoolDepositOp is an XDR Struct defines as:

LiquidityPoolEntry

LiquidityPoolEntry is an XDR Struct defines as:

LiquidityPoolEntryConstantProduct

LiquidityPoolEntryConstantProduct is an XDR NestedStruct defines as:

LiquidityPoolWithdrawOp

LiquidityPoolWithdrawOp is an XDR Struct defines as:

ManageBuyOfferOp

ManageBuyOfferOp is an XDR Struct defines as:

ManageDataOp

ManageDataOp is an XDR Struct defines as:

ManageOfferSuccessResult

ManageOfferSuccessResult is an XDR Struct defines as:

ManageSellOfferOp

ManageSellOfferOp is an XDR Struct defines as:

MuxedAccountMed25519

MuxedAccountMed25519 is an XDR NestedStruct defines as:

NodeId

NodeId is an XDR Typedef defines as:

OfferEntry

OfferEntry is an XDR Struct defines as:

Operation

Operation is an XDR Struct defines as:

OperationMeta

OperationMeta is an XDR Struct defines as:

PathPaymentStrictReceiveOp

PathPaymentStrictReceiveOp is an XDR Struct defines as:

PathPaymentStrictReceiveResultSuccess

PathPaymentStrictReceiveResultSuccess is an XDR NestedStruct defines as:

PathPaymentStrictSendOp

PathPaymentStrictSendOp is an XDR Struct defines as:

PathPaymentStrictSendResultSuccess

PathPaymentStrictSendResultSuccess is an XDR NestedStruct defines as:

PaymentOp

PaymentOp is an XDR Struct defines as:

PeerAddress

PeerAddress is an XDR Struct defines as:

PeerStatList

PeerStatList is an XDR Typedef defines as:

PeerStats

PeerStats is an XDR Struct defines as:

PersistedScpStateV0

PersistedScpStateV0 is an XDR Struct defines as:

PersistedScpStateV1

PersistedScpStateV1 is an XDR Struct defines as:

PoolId

PoolId is an XDR Typedef defines as:

PreconditionsV2

PreconditionsV2 is an XDR Struct defines as:

Price

Price is an XDR Struct defines as:

ReadXdrIter

RestoreFootprintOp

RestoreFootprintOp is an XDR Struct defines as:

RevokeSponsorshipOpSigner

RevokeSponsorshipOpSigner is an XDR NestedStruct defines as:

SError

SError is an XDR Struct defines as:

ScBytes

ScBytes is an XDR Typedef defines as:

ScContractInstance

ScContractInstance is an XDR Struct defines as:

ScMap

ScMap is an XDR Typedef defines as:

ScMapEntry

ScMapEntry is an XDR Struct defines as:

ScMetaV0

ScMetaV0 is an XDR Struct defines as:

ScNonceKey

ScNonceKey is an XDR Struct defines as:

ScSpecFunctionInputV0

ScSpecFunctionInputV0 is an XDR Struct defines as:

ScSpecFunctionV0

ScSpecFunctionV0 is an XDR Struct defines as:

ScSpecTypeBytesN

ScSpecTypeBytesN is an XDR Struct defines as:

ScSpecTypeMap

ScSpecTypeMap is an XDR Struct defines as:

ScSpecTypeOption

ScSpecTypeOption is an XDR Struct defines as:

ScSpecTypeResult

ScSpecTypeResult is an XDR Struct defines as:

ScSpecTypeTuple

ScSpecTypeTuple is an XDR Struct defines as:

ScSpecTypeUdt

ScSpecTypeUdt is an XDR Struct defines as:

ScSpecTypeVec

ScSpecTypeVec is an XDR Struct defines as:

ScSpecUdtEnumCaseV0

ScSpecUdtEnumCaseV0 is an XDR Struct defines as:

ScSpecUdtEnumV0

ScSpecUdtEnumV0 is an XDR Struct defines as:

ScSpecUdtErrorEnumCaseV0

ScSpecUdtErrorEnumCaseV0 is an XDR Struct defines as:

ScSpecUdtErrorEnumV0

ScSpecUdtErrorEnumV0 is an XDR Struct defines as:

ScSpecUdtStructFieldV0

ScSpecUdtStructFieldV0 is an XDR Struct defines as:

ScSpecUdtStructV0

ScSpecUdtStructV0 is an XDR Struct defines as:

ScSpecUdtUnionCaseTupleV0

ScSpecUdtUnionCaseTupleV0 is an XDR Struct defines as:

ScSpecUdtUnionCaseVoidV0

ScSpecUdtUnionCaseVoidV0 is an XDR Struct defines as:

ScSpecUdtUnionV0

ScSpecUdtUnionV0 is an XDR Struct defines as:

ScString

ScString is an XDR Typedef defines as:

ScSymbol

ScSymbol is an XDR Typedef defines as:

ScVec

ScVec is an XDR Typedef defines as:

ScpBallot

ScpBallot is an XDR Struct defines as:

ScpEnvelope

ScpEnvelope is an XDR Struct defines as:

ScpHistoryEntryV0

ScpHistoryEntryV0 is an XDR Struct defines as:

ScpNomination

ScpNomination is an XDR Struct defines as:

ScpQuorumSet

ScpQuorumSet is an XDR Struct defines as:

ScpStatement

ScpStatement is an XDR Struct defines as:

ScpStatementConfirm

ScpStatementConfirm is an XDR NestedStruct defines as:

ScpStatementExternalize

ScpStatementExternalize is an XDR NestedStruct defines as:

ScpStatementPrepare

ScpStatementPrepare is an XDR NestedStruct defines as:

SendMore

SendMore is an XDR Struct defines as:

SendMoreExtended

SendMoreExtended is an XDR Struct defines as:

SequenceNumber

SequenceNumber is an XDR Typedef defines as:

SetOptionsOp

SetOptionsOp is an XDR Struct defines as:

SetTrustLineFlagsOp

SetTrustLineFlagsOp is an XDR Struct defines as:

Signature

Signature is an XDR Typedef defines as:

SignatureHint

SignatureHint is an XDR Typedef defines as:

SignedSurveyRequestMessage

SignedSurveyRequestMessage is an XDR Struct defines as:

SignedSurveyResponseMessage

SignedSurveyResponseMessage is an XDR Struct defines as:

SignedTimeSlicedSurveyRequestMessage

SignedTimeSlicedSurveyRequestMessage is an XDR Struct defines as:

SignedTimeSlicedSurveyResponseMessage

SignedTimeSlicedSurveyResponseMessage is an XDR Struct defines as:

SignedTimeSlicedSurveyStartCollectingMessage

SignedTimeSlicedSurveyStartCollectingMessage is an XDR Struct defines as:

SignedTimeSlicedSurveyStopCollectingMessage

SignedTimeSlicedSurveyStopCollectingMessage is an XDR Struct defines as:

Signer

Signer is an XDR Struct defines as:

SignerKeyEd25519SignedPayload

SignerKeyEd25519SignedPayload is an XDR NestedStruct defines as:

SimplePaymentResult

SimplePaymentResult is an XDR Struct defines as:

SorobanAddressCredentials

SorobanAddressCredentials is an XDR Struct defines as:

SorobanAuthorizationEntry

SorobanAuthorizationEntry is an XDR Struct defines as:

SorobanAuthorizedInvocation

SorobanAuthorizedInvocation is an XDR Struct defines as:

SorobanResources

SorobanResources is an XDR Struct defines as:

SorobanTransactionData

SorobanTransactionData is an XDR Struct defines as:

SorobanTransactionMeta

SorobanTransactionMeta is an XDR Struct defines as:

SorobanTransactionMetaExtV1

SorobanTransactionMetaExtV1 is an XDR Struct defines as:

SponsorshipDescriptor

SponsorshipDescriptor is an XDR Typedef defines as:

StateArchivalSettings

StateArchivalSettings is an XDR Struct defines as:

StellarValue

StellarValue is an XDR Struct defines as:

StoredDebugTransactionSet

StoredDebugTransactionSet is an XDR Struct defines as:

String32

String32 is an XDR Typedef defines as:

String64

String64 is an XDR Typedef defines as:

StringM

A string type that contains arbitrary bytes.

SurveyRequestMessage

SurveyRequestMessage is an XDR Struct defines as:

SurveyResponseMessage

SurveyResponseMessage is an XDR Struct defines as:

Thresholds

Thresholds is an XDR Typedef defines as:

TimeBounds

TimeBounds is an XDR Struct defines as:

TimePoint

TimePoint is an XDR Typedef defines as:

TimeSlicedNodeData

TimeSlicedNodeData is an XDR Struct defines as:

TimeSlicedPeerData

TimeSlicedPeerData is an XDR Struct defines as:

TimeSlicedPeerDataList

TimeSlicedPeerDataList is an XDR Typedef defines as:

TimeSlicedSurveyRequestMessage

TimeSlicedSurveyRequestMessage is an XDR Struct defines as:

TimeSlicedSurveyResponseMessage

TimeSlicedSurveyResponseMessage is an XDR Struct defines as:

TimeSlicedSurveyStartCollectingMessage

TimeSlicedSurveyStartCollectingMessage is an XDR Struct defines as:

TimeSlicedSurveyStopCollectingMessage

TimeSlicedSurveyStopCollectingMessage is an XDR Struct defines as:

TopologyResponseBodyV0

TopologyResponseBodyV0 is an XDR Struct defines as:

TopologyResponseBodyV1

TopologyResponseBodyV1 is an XDR Struct defines as:

TopologyResponseBodyV2

TopologyResponseBodyV2 is an XDR Struct defines as:

Transaction

Transaction is an XDR Struct defines as:

TransactionHistoryEntry

TransactionHistoryEntry is an XDR Struct defines as:

TransactionHistoryResultEntry

TransactionHistoryResultEntry is an XDR Struct defines as:

TransactionMetaV1

TransactionMetaV1 is an XDR Struct defines as:

TransactionMetaV2

TransactionMetaV2 is an XDR Struct defines as:

TransactionMetaV3

TransactionMetaV3 is an XDR Struct defines as:

TransactionResult

TransactionResult is an XDR Struct defines as:

TransactionResultMeta

TransactionResultMeta is an XDR Struct defines as:

TransactionResultPair

TransactionResultPair is an XDR Struct defines as:

TransactionResultSet

TransactionResultSet is an XDR Struct defines as:

TransactionSet

TransactionSet is an XDR Struct defines as:

TransactionSetV1

TransactionSetV1 is an XDR Struct defines as:

TransactionSignaturePayload

TransactionSignaturePayload is an XDR Struct defines as:

TransactionV0

TransactionV0 is an XDR Struct defines as:

TransactionV0Envelope

TransactionV0Envelope is an XDR Struct defines as:

TransactionV1Envelope

TransactionV1Envelope is an XDR Struct defines as:

TrustLineEntry

TrustLineEntry is an XDR Struct defines as:

TrustLineEntryExtensionV2

TrustLineEntryExtensionV2 is an XDR Struct defines as:

TrustLineEntryV1

TrustLineEntryV1 is an XDR NestedStruct defines as:

TtlEntry

TtlEntry is an XDR Struct defines as:

TxAdvertVector

TxAdvertVector is an XDR Typedef defines as:

TxDemandVector

TxDemandVector is an XDR Typedef defines as:

TxSetComponentTxsMaybeDiscountedFee

TxSetComponentTxsMaybeDiscountedFee is an XDR NestedStruct defines as:

UInt128Parts

UInt128Parts is an XDR Struct defines as:

UInt256Parts

UInt256Parts is an XDR Struct defines as:

Uint256

Uint256 is an XDR Typedef defines as:

UpgradeEntryMeta

UpgradeEntryMeta is an XDR Struct defines as:

UpgradeType

UpgradeType is an XDR Typedef defines as:

Value

Value is an XDR Typedef defines as:

VecM

Enums

AccountEntryExt

AccountEntryExt is an XDR NestedUnion defines as:

AccountEntryExtensionV1Ext

AccountEntryExtensionV1Ext is an XDR NestedUnion defines as:

AccountEntryExtensionV2Ext

AccountEntryExtensionV2Ext is an XDR NestedUnion defines as:

AccountFlags

AccountFlags is an XDR Enum defines as:

AccountMergeResult

AccountMergeResult is an XDR Union defines as:

AccountMergeResultCode

AccountMergeResultCode is an XDR Enum defines as:

AllowTrustResult

AllowTrustResult is an XDR Union defines as:

AllowTrustResultCode

AllowTrustResultCode is an XDR Enum defines as:

Asset

Asset is an XDR Union defines as:

AssetCode

AssetCode is an XDR Union defines as:

AssetType

AssetType is an XDR Enum defines as:

AuthenticatedMessage

AuthenticatedMessage is an XDR Union defines as:

BeginSponsoringFutureReservesResult

BeginSponsoringFutureReservesResult is an XDR Union defines as:

BeginSponsoringFutureReservesResultCode

BeginSponsoringFutureReservesResultCode is an XDR Enum defines as:

BucketEntry

BucketEntry is an XDR Union defines as:

BucketEntryType

BucketEntryType is an XDR Enum defines as:

BucketMetadataExt

BucketMetadataExt is an XDR NestedUnion defines as:

BumpSequenceResult

BumpSequenceResult is an XDR Union defines as:

BumpSequenceResultCode

BumpSequenceResultCode is an XDR Enum defines as:

ChangeTrustAsset

ChangeTrustAsset is an XDR Union defines as:

ChangeTrustResult

ChangeTrustResult is an XDR Union defines as:

ChangeTrustResultCode

ChangeTrustResultCode is an XDR Enum defines as:

ClaimAtom

ClaimAtom is an XDR Union defines as:

ClaimAtomType

ClaimAtomType is an XDR Enum defines as:

ClaimClaimableBalanceResult

ClaimClaimableBalanceResult is an XDR Union defines as:

ClaimClaimableBalanceResultCode

ClaimClaimableBalanceResultCode is an XDR Enum defines as:

ClaimPredicate

ClaimPredicate is an XDR Union defines as:

ClaimPredicateType

ClaimPredicateType is an XDR Enum defines as:

ClaimableBalanceEntryExt

ClaimableBalanceEntryExt is an XDR NestedUnion defines as:

ClaimableBalanceEntryExtensionV1Ext

ClaimableBalanceEntryExtensionV1Ext is an XDR NestedUnion defines as:

ClaimableBalanceFlags

ClaimableBalanceFlags is an XDR Enum defines as:

ClaimableBalanceId

ClaimableBalanceId is an XDR Union defines as:

ClaimableBalanceIdType

ClaimableBalanceIdType is an XDR Enum defines as:

Claimant

Claimant is an XDR Union defines as:

ClaimantType

ClaimantType is an XDR Enum defines as:

ClawbackClaimableBalanceResult

ClawbackClaimableBalanceResult is an XDR Union defines as:

ClawbackClaimableBalanceResultCode

ClawbackClaimableBalanceResultCode is an XDR Enum defines as:

ClawbackResult

ClawbackResult is an XDR Union defines as:

ClawbackResultCode

ClawbackResultCode is an XDR Enum defines as:

ConfigSettingEntry

ConfigSettingEntry is an XDR Union defines as:

ConfigSettingId

ConfigSettingId is an XDR Enum defines as:

ContractCodeEntryExt

ContractCodeEntryExt is an XDR NestedUnion defines as:

ContractCostType

ContractCostType is an XDR Enum defines as:

ContractDataDurability

ContractDataDurability is an XDR Enum defines as:

ContractEventBody

ContractEventBody is an XDR NestedUnion defines as:

ContractEventType

ContractEventType is an XDR Enum defines as:

ContractExecutable

ContractExecutable is an XDR Union defines as:

ContractExecutableType

ContractExecutableType is an XDR Enum defines as:

ContractIdPreimage

ContractIdPreimage is an XDR Union defines as:

ContractIdPreimageType

ContractIdPreimageType is an XDR Enum defines as:

CreateAccountResult

CreateAccountResult is an XDR Union defines as:

CreateAccountResultCode

CreateAccountResultCode is an XDR Enum defines as:

CreateClaimableBalanceResult

CreateClaimableBalanceResult is an XDR Union defines as:

CreateClaimableBalanceResultCode

CreateClaimableBalanceResultCode is an XDR Enum defines as:

CryptoKeyType

CryptoKeyType is an XDR Enum defines as:

DataEntryExt

DataEntryExt is an XDR NestedUnion defines as:

EndSponsoringFutureReservesResult

EndSponsoringFutureReservesResult is an XDR Union defines as:

EndSponsoringFutureReservesResultCode

EndSponsoringFutureReservesResultCode is an XDR Enum defines as:

EnvelopeType

EnvelopeType is an XDR Enum defines as:

Error

Error contains all errors returned by functions in this crate. It can be compared via PartialEq, however any contained IO errors will only be compared on their ErrorKind.

ErrorCode

ErrorCode is an XDR Enum defines as:

ExtendFootprintTtlResult

ExtendFootprintTtlResult is an XDR Union defines as:

ExtendFootprintTtlResultCode

ExtendFootprintTtlResultCode is an XDR Enum defines as:

ExtensionPoint

ExtensionPoint is an XDR Union defines as:

FeeBumpTransactionExt

FeeBumpTransactionExt is an XDR NestedUnion defines as:

FeeBumpTransactionInnerTx

FeeBumpTransactionInnerTx is an XDR NestedUnion defines as:

GeneralizedTransactionSet

GeneralizedTransactionSet is an XDR Union defines as:

HashIdPreimage

HashIdPreimage is an XDR Union defines as:

HostFunction

HostFunction is an XDR Union defines as:

HostFunctionType

HostFunctionType is an XDR Enum defines as:

InflationResult

InflationResult is an XDR Union defines as:

InflationResultCode

InflationResultCode is an XDR Enum defines as:

InnerTransactionResultExt

InnerTransactionResultExt is an XDR NestedUnion defines as:

InnerTransactionResultResult

InnerTransactionResultResult is an XDR NestedUnion defines as:

InvokeHostFunctionResult

InvokeHostFunctionResult is an XDR Union defines as:

InvokeHostFunctionResultCode

InvokeHostFunctionResultCode is an XDR Enum defines as:

IpAddrType

IpAddrType is an XDR Enum defines as:

LedgerCloseMeta

LedgerCloseMeta is an XDR Union defines as:

LedgerCloseMetaExt

LedgerCloseMetaExt is an XDR Union defines as:

LedgerEntryChange

LedgerEntryChange is an XDR Union defines as:

LedgerEntryChangeType

LedgerEntryChangeType is an XDR Enum defines as:

LedgerEntryData

LedgerEntryData is an XDR NestedUnion defines as:

LedgerEntryExt

LedgerEntryExt is an XDR NestedUnion defines as:

LedgerEntryExtensionV1Ext

LedgerEntryExtensionV1Ext is an XDR NestedUnion defines as:

LedgerEntryType

LedgerEntryType is an XDR Enum defines as:

LedgerHeaderExt

LedgerHeaderExt is an XDR NestedUnion defines as:

LedgerHeaderExtensionV1Ext

LedgerHeaderExtensionV1Ext is an XDR NestedUnion defines as:

LedgerHeaderFlags

LedgerHeaderFlags is an XDR Enum defines as:

LedgerHeaderHistoryEntryExt

LedgerHeaderHistoryEntryExt is an XDR NestedUnion defines as:

LedgerKey

LedgerKey is an XDR Union defines as:

LedgerUpgrade

LedgerUpgrade is an XDR Union defines as:

LedgerUpgradeType

LedgerUpgradeType is an XDR Enum defines as:

LiquidityPoolDepositResult

LiquidityPoolDepositResult is an XDR Union defines as:

LiquidityPoolDepositResultCode

LiquidityPoolDepositResultCode is an XDR Enum defines as:

LiquidityPoolEntryBody

LiquidityPoolEntryBody is an XDR NestedUnion defines as:

LiquidityPoolParameters

LiquidityPoolParameters is an XDR Union defines as:

LiquidityPoolType

LiquidityPoolType is an XDR Enum defines as:

LiquidityPoolWithdrawResult

LiquidityPoolWithdrawResult is an XDR Union defines as:

LiquidityPoolWithdrawResultCode

LiquidityPoolWithdrawResultCode is an XDR Enum defines as:

ManageBuyOfferResult

ManageBuyOfferResult is an XDR Union defines as:

ManageBuyOfferResultCode

ManageBuyOfferResultCode is an XDR Enum defines as:

ManageDataResult

ManageDataResult is an XDR Union defines as:

ManageDataResultCode

ManageDataResultCode is an XDR Enum defines as:

ManageOfferEffect

ManageOfferEffect is an XDR Enum defines as:

ManageOfferSuccessResultOffer

ManageOfferSuccessResultOffer is an XDR NestedUnion defines as:

ManageSellOfferResult

ManageSellOfferResult is an XDR Union defines as:

ManageSellOfferResultCode

ManageSellOfferResultCode is an XDR Enum defines as:

Memo

Memo is an XDR Union defines as:

MemoType

MemoType is an XDR Enum defines as:

MessageType

MessageType is an XDR Enum defines as:

MuxedAccount

MuxedAccount is an XDR Union defines as:

OfferEntryExt

OfferEntryExt is an XDR NestedUnion defines as:

OfferEntryFlags

OfferEntryFlags is an XDR Enum defines as:

OperationBody

OperationBody is an XDR NestedUnion defines as:

OperationResult

OperationResult is an XDR Union defines as:

OperationResultCode

OperationResultCode is an XDR Enum defines as:

OperationResultTr

OperationResultTr is an XDR NestedUnion defines as:

OperationType

OperationType is an XDR Enum defines as:

PathPaymentStrictReceiveResult

PathPaymentStrictReceiveResult is an XDR Union defines as:

PathPaymentStrictReceiveResultCode

PathPaymentStrictReceiveResultCode is an XDR Enum defines as:

PathPaymentStrictSendResult

PathPaymentStrictSendResult is an XDR Union defines as:

PathPaymentStrictSendResultCode

PathPaymentStrictSendResultCode is an XDR Enum defines as:

PaymentResult

PaymentResult is an XDR Union defines as:

PaymentResultCode

PaymentResultCode is an XDR Enum defines as:

PeerAddressIp

PeerAddressIp is an XDR NestedUnion defines as:

PersistedScpState

PersistedScpState is an XDR Union defines as:

PreconditionType

PreconditionType is an XDR Enum defines as:

Preconditions

Preconditions is an XDR Union defines as:

PublicKey

PublicKey is an XDR Union defines as:

PublicKeyType

PublicKeyType is an XDR Enum defines as:

RestoreFootprintResult

RestoreFootprintResult is an XDR Union defines as:

RestoreFootprintResultCode

RestoreFootprintResultCode is an XDR Enum defines as:

RevokeSponsorshipOp

RevokeSponsorshipOp is an XDR Union defines as:

RevokeSponsorshipResult

RevokeSponsorshipResult is an XDR Union defines as:

RevokeSponsorshipResultCode

RevokeSponsorshipResultCode is an XDR Enum defines as:

RevokeSponsorshipType

RevokeSponsorshipType is an XDR Enum defines as:

ScAddress

ScAddress is an XDR Union defines as:

ScAddressType

ScAddressType is an XDR Enum defines as:

ScEnvMetaEntry

ScEnvMetaEntry is an XDR Union defines as:

ScEnvMetaKind

ScEnvMetaKind is an XDR Enum defines as:

ScError

ScError is an XDR Union defines as:

ScErrorCode

ScErrorCode is an XDR Enum defines as:

ScErrorType

ScErrorType is an XDR Enum defines as:

ScMetaEntry

ScMetaEntry is an XDR Union defines as:

ScMetaKind

ScMetaKind is an XDR Enum defines as:

ScSpecEntry

ScSpecEntry is an XDR Union defines as:

ScSpecEntryKind

ScSpecEntryKind is an XDR Enum defines as:

ScSpecType

ScSpecType is an XDR Enum defines as:

ScSpecTypeDef

ScSpecTypeDef is an XDR Union defines as:

ScSpecUdtUnionCaseV0

ScSpecUdtUnionCaseV0 is an XDR Union defines as:

ScSpecUdtUnionCaseV0Kind

ScSpecUdtUnionCaseV0Kind is an XDR Enum defines as:

ScVal

ScVal is an XDR Union defines as:

ScValType

ScValType is an XDR Enum defines as:

ScpHistoryEntry

ScpHistoryEntry is an XDR Union defines as:

ScpStatementPledges

ScpStatementPledges is an XDR NestedUnion defines as:

ScpStatementType

ScpStatementType is an XDR Enum defines as:

SetOptionsResult

SetOptionsResult is an XDR Union defines as:

SetOptionsResultCode

SetOptionsResultCode is an XDR Enum defines as:

SetTrustLineFlagsResult

SetTrustLineFlagsResult is an XDR Union defines as:

SetTrustLineFlagsResultCode

SetTrustLineFlagsResultCode is an XDR Enum defines as:

SignerKey

SignerKey is an XDR Union defines as:

SignerKeyType

SignerKeyType is an XDR Enum defines as:

SorobanAuthorizedFunction

SorobanAuthorizedFunction is an XDR Union defines as:

SorobanAuthorizedFunctionType

SorobanAuthorizedFunctionType is an XDR Enum defines as:

SorobanCredentials

SorobanCredentials is an XDR Union defines as:

SorobanCredentialsType

SorobanCredentialsType is an XDR Enum defines as:

SorobanTransactionMetaExt

SorobanTransactionMetaExt is an XDR Union defines as:

StellarMessage

StellarMessage is an XDR Union defines as:

StellarValueExt

StellarValueExt is an XDR NestedUnion defines as:

StellarValueType

StellarValueType is an XDR Enum defines as:

StoredTransactionSet

StoredTransactionSet is an XDR Union defines as:

SurveyMessageCommandType

SurveyMessageCommandType is an XDR Enum defines as:

SurveyMessageResponseType

SurveyMessageResponseType is an XDR Enum defines as:

SurveyResponseBody

SurveyResponseBody is an XDR Union defines as:

ThresholdIndexes

ThresholdIndexes is an XDR Enum defines as:

TransactionEnvelope

TransactionEnvelope is an XDR Union defines as:

TransactionExt

TransactionExt is an XDR NestedUnion defines as:

TransactionHistoryEntryExt

TransactionHistoryEntryExt is an XDR NestedUnion defines as:

TransactionHistoryResultEntryExt

TransactionHistoryResultEntryExt is an XDR NestedUnion defines as:

TransactionMeta

TransactionMeta is an XDR Union defines as:

TransactionPhase

TransactionPhase is an XDR Union defines as:

TransactionResultCode

TransactionResultCode is an XDR Enum defines as:

TransactionResultExt

TransactionResultExt is an XDR NestedUnion defines as:

TransactionResultResult

TransactionResultResult is an XDR NestedUnion defines as:

TransactionSignaturePayloadTaggedTransaction

TransactionSignaturePayloadTaggedTransaction is an XDR NestedUnion defines as:

TransactionV0Ext

TransactionV0Ext is an XDR NestedUnion defines as:

TrustLineAsset

TrustLineAsset is an XDR Union defines as:

TrustLineEntryExt

TrustLineEntryExt is an XDR NestedUnion defines as:

TrustLineEntryExtensionV2Ext

TrustLineEntryExtensionV2Ext is an XDR NestedUnion defines as:

TrustLineEntryV1Ext

TrustLineEntryV1Ext is an XDR NestedUnion defines as:

TrustLineFlags

TrustLineFlags is an XDR Enum defines as:

TxSetComponent

TxSetComponent is an XDR Union defines as:

TxSetComponentType

TxSetComponentType is an XDR Enum defines as:

Type

TypeVariant

Constants

AUTH_MSG_FLAG_FLOW_CONTROL_BYTES_REQUESTED

AuthMsgFlagFlowControlBytesRequested is an XDR Const defines as:

CONTRACT_COST_COUNT_LIMIT

ContractCostCountLimit is an XDR Const defines as:

LIQUIDITY_POOL_FEE_V18

LiquidityPoolFeeV18 is an XDR Const defines as:

MASK_ACCOUNT_FLAGS

MaskAccountFlags is an XDR Const defines as:

MASK_ACCOUNT_FLAGS_V17

MaskAccountFlagsV17 is an XDR Const defines as:

MASK_CLAIMABLE_BALANCE_FLAGS

MaskClaimableBalanceFlags is an XDR Const defines as:

MASK_LEDGER_HEADER_FLAGS

MaskLedgerHeaderFlags is an XDR Const defines as:

MASK_OFFERENTRY_FLAGS

MaskOfferentryFlags is an XDR Const defines as:

MASK_TRUSTLINE_FLAGS

MaskTrustlineFlags is an XDR Const defines as:

MASK_TRUSTLINE_FLAGS_V13

MaskTrustlineFlagsV13 is an XDR Const defines as:

MASK_TRUSTLINE_FLAGS_V17

MaskTrustlineFlagsV17 is an XDR Const defines as:

MAX_OPS_PER_TX

MaxOpsPerTx is an XDR Const defines as:

MAX_SIGNERS

MaxSigners is an XDR Const defines as:

SCSYMBOL_LIMIT

ScsymbolLimit is an XDR Const defines as:

SC_SPEC_DOC_LIMIT

ScSpecDocLimit is an XDR Const defines as:

TX_ADVERT_VECTOR_MAX_SIZE

TxAdvertVectorMaxSize is an XDR Const defines as:

TX_DEMAND_VECTOR_MAX_SIZE

TxDemandVectorMaxSize is an XDR Const defines as:

XDR_FILES_SHA256

XDR_FILES_SHA256 is a list of pairs of source files and their SHA256 hashes.

Traits

Discriminant

Discriminant defines types that may contain a one-of value determined according to the discriminant, and exposes the value of the discriminant for that type, such as in an XDR union.

Enum

Name

Name defines types that assign a static name to their value, such as the name given to an identifier in an XDR enum, or the name given to the case in a union.

ReadXdr

Union

Validate

Variants

Iter defines types that have variants that can be iterated.

WriteXdr

Type Aliases

Int32

Int32 is an XDR Typedef defines as:

Int64

Int64 is an XDR Typedef defines as:

Uint32

Uint32 is an XDR Typedef defines as:

Uint64

Uint64 is an XDR Typedef defines as: