soroban_env_common
Trait CheckedEnvArg

pub trait CheckedEnvArg: Sized {
    // Provided method
    fn check_env_arg<E: Env>(self, _e: &E) -> Result<Self, E::Error> { ... }
}

This trait is used by macro-generated dispatch and forwarding functions to check arguments being passed to the Env. The default implementations call through to the Env integrity-checking functions.

Provided Methods

fn check_env_arg<E: Env>(self, _e: &E) -> Result<Self, E::Error>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.